requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
google-cloud-firestore>=2.13.0
openai>=1.0.0
python-multipart
PyMuPDF==1.26.4
numpy>=1.26.0

requests>=2.31.0
pdfplumber>=0.11.0
google-cloud-storage>=2.16.0

openpyxl>=3.1.2
python-docx>=1.1.2
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
